Missionparse work for objects
Very much a work in progress. I have tried to comment as much as
possible to make my thinking clearer.

When the asteroids are tabled in FotG we can point the cases to them.
The closest mine is currently the ION mine but the weapon can be changed
in FRED.

As for the backgrounds. It would probably be best to not process these
at all from the game and perhaps at a later stage, add FotG backgrounds
depending on where the location of the mission is - from the briefing.

For the mine field, because it is in a 2d plane with each mine having
coordinates (a,b) I felt the best way was to have two for loops... the
first would move one step along one axis, then the inner for loop would
populate the mines along the perpendicular axis. The grid is populated
starting on the initial mine as the origin however in XWing the mine is
in the centre of the grid. Therefore the second part of the calculation
is to shift the mine back so centre the grid on the given position.

This second part of the calculation could be done before the for loops
which means it would only have to be done once and would also help make
the code clearer.