Separate the function to check for unique suffix
For the unique name check for the object, put this into a separate
function. Update missionparse to add default values for the object
initialisation.

Add a goto for now in case the search for a unique name gets stuck in an
infinte loop. This can potentially be removed in a later PR.