FRED import fixes
This is the equivalent of scp-fs2open#5776 for the X-Wing importer.  There's also a touch of scp-fs2open#5377 because the `OnFileImportFSM` changes from that PR were not carried over to `OnFileImportXWI`.  Finally, there is an additional try/catch block when parsing the .BRF file because we don't want the import to fail if the briefing is missing.