flag +nocreate ships to check that they have sources
Print a warning for all ships in modular tables that have +nocreate and are missing the source ship, in case the source ship is not actually supposed to be missing.  There is a boolean flag to skip through all warnings remaining in a file in case the file does not need to be checked.